move `EthereumRoutingModel` and `EthereumTrader` base classes
