+++ tradeexecutor/ethereum/uniswap_v3_routing.py
@@ -23,7 +23,7 @@ from tradeexecutor.ethereum.routing_state import (
     route_tokens, # don't remove, forwarded import
     OutOfBalance, # don't remove, forwarded import
 )
-from tradeexecutor.strategy.routing_model import EthereumRoutingModel
+from tradeexecutor.ethereum.routing_model import EthereumRoutingModel
 
 logger = logging.getLogger(__name__)
